Scope current courses controller under providers module
