+++ config/routes.rb
@@ -102,8 +102,10 @@ Rails.application.routes.draw do
     namespace :public do
       namespace :v1 do
         resources :recruitment_cycles, param: :year, only: [:show] do
-          resources :providers, only: %i[index show], param: :code do
-            resources :courses, only: %i[index show] do
+          resources :courses, only: %i[index]
+
+          resources :providers, only: %i[index show] do
+            resources :courses, only: %i[index show], module: :providers do
               resources :locations, only: [:index]
             end
           end
