+++ spec/api/providers/courses_spec.rb
